+++ llm_guard/input_scanners/ban_topics.py
@@ -8,13 +8,13 @@ from .base import Scanner
 
 LOGGER = get_logger()
 
-# This model was trained on a mixture of 33 datasets and 389 classes reformatted in the universal NLI format.
-# The model is English only. You can also use it for multilingual zeroshot classification by first machine translating texts to English.
-MODEL_LARGE = Model(
-    path="MoritzLaurer/deberta-v3-large-zeroshot-v1.1-all-33",
-    revision="c5dca3bda16d30337e493e3e3e5caa19a3e7c8c2",
-    onnx_path="MoritzLaurer/deberta-v3-large-zeroshot-v1.1-all-33",
-    onnx_revision="c5dca3bda16d30337e493e3e3e5caa19a3e7c8c2",
+# The most performant model. 0.43 B parameters, 870 MB.
+# It's English only. Context length max 512 tokens
+MODEL_DEBERTA_LARGE_V2 = Model(
+    path="MoritzLaurer/deberta-v3-large-zeroshot-v2.0",
+    revision="cf44676c28ba7312e5c5f8f8d2c22b3e0c9cdae2",
+    onnx_path="MoritzLaurer/deberta-v3-large-zeroshot-v2.0",
+    onnx_revision="cf44676c28ba7312e5c5f8f8d2c22b3e0c9cdae2",
     onnx_subfolder="onnx",
     pipeline_kwargs={
         "max_length": 512,
@@ -22,27 +22,56 @@ MODEL_LARGE = Model(
     },
 )
 
-# This is essentially the same as its larger sister MoritzLaurer/deberta-v3-large-zeroshot-v1.1-all-33 only that it's smaller.
-# Use it if you need more speed. The model is English-only.
-MODEL_BASE = Model(
-    path="MoritzLaurer/deberta-v3-base-zeroshot-v1.1-all-33",
-    revision="7520e29795fdb58f884ef65ef6c46de2a5a8a63c",
-    onnx_path="MoritzLaurer/deberta-v3-base-zeroshot-v1.1-all-33",
-    onnx_revision="7520e29795fdb58f884ef65ef6c46de2a5a8a63c",
+# The most performant base model. 0.18 B parameters, 369 MB.
+# It's English only. Context length max 512 tokens.
+# Faster than RoBERTa-large/BGE-3 models, but slower than RoBERTa-base
+MODEL_DEBERTA_BASE_V2 = Model(
+    path="MoritzLaurer/deberta-v3-base-zeroshot-v2.0",
+    revision="8e7e5af5983a0ddb1a5b45a38b129ab69e2258e8",
+    onnx_path="MoritzLaurer/deberta-v3-base-zeroshot-v2.0",
     onnx_subfolder="onnx",
+    onnx_revision="8e7e5af5983a0ddb1a5b45a38b129ab69e2258e8",
     pipeline_kwargs={
         "max_length": 512,
         "truncation": True,
     },
 )
 
-# Same as above, just smaller/faster.
-MODEL_XSMALL = Model(
-    path="MoritzLaurer/deberta-v3-xsmall-zeroshot-v1.1-all-33",
-    revision="02d678b25107a8e8e9d41276a593a4b43ebcf9ca",
-    onnx_path="MoritzLaurer/deberta-v3-xsmall-zeroshot-v1.1-all-33",
-    onnx_revision="02d678b25107a8e8e9d41276a593a4b43ebcf9ca",
+# The most performance multilingual model. 0.57 B parameters, 1.14 GB.
+# 100+ languages; context length max 8192 tokens; based on bge-m3-retromae, which is based on XLM-RoBERTa
+MODEL_BGE_M3_V2 = Model(
+    path="MoritzLaurer/bge-m3-zeroshot-v2.0",
+    revision="cd3f8598c7359a3b5cbce164d7fcdafb83a36484",
+    onnx_path="MoritzLaurer/bge-m3-zeroshot-v2.0",
     onnx_subfolder="onnx",
+    onnx_revision="cd3f8598c7359a3b5cbce164d7fcdafb83a36484",
+    pipeline_kwargs={
+        "max_length": 8192,
+        "truncation": True,
+    },
+)
+
+# Less performant than deberta-v3 variants, but a bit faster and compatible with flash attention and TEI containers.
+# Size: 0.35B parameters, 711 MB. It's English only. Context length max 512 tokens.
+# Only trained on commercially-friendly data.
+MODEL_ROBERTA_LARGE_C_V2 = Model(
+    path="MoritzLaurer/roberta-large-zeroshot-v2.0-c",
+    revision="4c24ed4bba5af8d3162604abc2a141b9d2183ecc",
+    onnx_path="MoritzLaurer/roberta-large-zeroshot-v2.0-c",
+    onnx_subfolder="onnx",
+    onnx_revision="4c24ed4bba5af8d3162604abc2a141b9d2183ecc",
+    pipeline_kwargs={
+        "max_length": 512,
+        "truncation": True,
+    },
+)
+
+# Same model but smaller, more efficient version.
+MODEL_ROBERTA_BASE_C_V2 = Model(
+    path="MoritzLaurer/roberta-base-zeroshot-v2.0-c",
+    revision="d825e740e0c59881cf0b0b1481ccf726b6d65341",
+    onnx_path="protectai/MoritzLaurer-roberta-base-zeroshot-v2.0-c-onnx",
+    onnx_revision="fde5343dbad32f1a5470890505c72ec656db6dbe",
     pipeline_kwargs={
         "max_length": 512,
         "truncation": True,
@@ -81,7 +110,7 @@ class BanTopics(Scanner):
         self._threshold = threshold
 
         if model is None:
-            model = MODEL_BASE
+            model = MODEL_DEBERTA_BASE_V2
 
         tf_tokenizer, tf_model = get_tokenizer_and_model_for_classification(
             model=model,